+++ pysui/sui/sui_pgql/pgql_schema.py
@@ -95,7 +95,7 @@ def load_schema_cache(
     """."""
     schema_mgr: Schema = None
     # TODO: Temporary Sui devnet bug workaround
-    if genv != "devnet":
+    if not genv.startswith("devnet"):
         _furl = gurl + "/" + default_version if default_version else gurl
     else:
         _furl = gurl
@@ -131,7 +131,7 @@ def load_schema_cache(
             _url, _init_client, _base_version, _long_version, _schema, _rpc_config
         )
 
-    if genv == "devnet":
+    if genv.startswith("devnet"):
         return schema_mgr
 
     while schema_mgr._present_schemas: