+++ pysui/sui/sui_txn/transaction.py
@@ -389,12 +389,6 @@ class _SuiTransactionBase:
                 case _:
                     raise ValueError(f"Uknown class type handler {clz_name}")
 
-    def _to_bytes_from_str(self, inbound: Union[str, SuiString]) -> list[int]:
-        """Utility to convert base64 string to bytes then as list of u8."""
-        return list(
-            base64.b64decode(inbound if isinstance(inbound, str) else inbound.value)
-        )
-
     @versionchanged(
         version="0.50.0",
         reason="Removed with_unpublished_dependencies and skip_fetch_latest_git_deps replace with `args_list`",
@@ -411,13 +405,12 @@ class _SuiTransactionBase:
             src_path,
             args_list,
         )
-        modules = list(map(self._to_bytes_from_str, compiled_package.compiled_modules))
         dependencies = [
             bcs.Address.from_str(x if isinstance(x, str) else x.value)
             for x in compiled_package.dependencies
         ]
         digest = bcs.Digest.from_bytes(compiled_package.package_digest)
-        return modules, dependencies, digest
+        return compiled_package.compiled_modules, dependencies, digest
 
     def _receiving_parm(self, parm) -> bool:
         """."""