+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.66.0] - Unpublished
+## [0.66.0] - 2024-08-04
 
 ### Added
 
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Coin deserialization failed when previousTransactionBlock has been pruned
 - Validate GraphQL schema existence before adding to available list
+- [bug](https://github.com/FrankC01/pysui/issues/222) Fixed publish that broke with enums
 
 ### Changed